MQTT-decryptor/requirements.txt
meshcore
python-dotenv
paho-mqtt
